// src-ui/core/ipc/generated/domain/RemoteSelectedPaths.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * 启动与组件工厂实际使用的路径
 */
export type RemoteSelectedPaths = { home: string, qqInstallBase?: string | null, qqBin?: string | null, napcatRoot?: string | null, snowlumaDir?: string | null, snowlumaWorkspace?: string | null, nodeBin?: string | null, ncdWatchRoot?: string | null, needsSudo: boolean, };